+++ src/test/index.ts
@@ -1,3 +1,4 @@
+export * from "./assert-view-state.js";
 export * from "./poll.js";
 export * from "./queue.js";
 export * from "./run-async.js";